refactor(portfolio): split explicit installation phases
