+++ tests/integration/sat_smt/test_smt_unsat_proof_verification.py
@@ -100,7 +100,7 @@ def _runtime_with_runtime(
     runtime = carcara_provider_runtime(executable)
     assert runtime.availability is CapabilityProviderAvailability.AVAILABLE
     monkeypatch.setattr(
-        "jacobian.portfolio.assembler.carcara_provider_runtime",
+        "jacobian.portfolio.provider_resolution.carcara_provider_runtime",
         lambda *_args, **_kwargs: runtime,
     )
     return create_runtime(
@@ -261,7 +261,7 @@ def test_proof_verify_requires_runtime_and_operator_authorization(
     )
     unavailable = carcara_provider_runtime(tmp_path / "missing")
     monkeypatch.setattr(
-        "jacobian.portfolio.assembler.carcara_provider_runtime",
+        "jacobian.portfolio.provider_resolution.carcara_provider_runtime",
         lambda *_args, **_kwargs: unavailable,
     )
     without_runtime = create_runtime(
